backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
alembic==1.14.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via starlette
ecdsa==0.19.0
    # via python-jose
fastapi==0.115.11
    # via -r requirements.in
greenlet==3.1.1
    # via sqlalchemy
idna==3.10
    # via anyio
joblib==1.4.2
    # via -r requirements.in
jose==1.0.0
    # via -r requirements.in
mako==1.3.9
    # via alembic
markupsafe==3.0.2
    # via mako
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pydantic==2.10.6
    # via
    #   -r requirements.in
    #   fastapi
pydantic-core==2.27.2
    # via pydantic
bcrypt==4.2.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
python-jose==3.4.0
    # via -r requirements.in
rsa==4.9
    # via python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.38
    # via
    #   -r requirements.in
    #   alembic
starlette==0.46.0
    # via fastapi
typing-extensions==4.12.2
    # via
    #   alembic
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn==0.34.0
psycopg2-binary==2.9.10
python-multipart==0.0.20
scikit-learn==1.5.0
scipy==1.15.2
numpy==2.2.3
cloudinary==1.43.0